Add files via upload
Updated midterm code, including building the model. While I was doing that, I found a new model with a better MSE (also better without outliers, i did the test but it's not in there yet) so I don't know if you guys think we should switch to this new one. Also, there are a few things i haven't had the time to test yet, like interaction terms with this new model.